refactor: remove unwanted exports from src/index.ts
